+++ app/Http/Controllers/Admin/Setting/AboutUsSettingController.php
@@ -6,30 +6,31 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\Admin\Settings\AdminAboutUsSettingUpdateRequest;
 use App\Models\AboutUsSetting;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\File;
 
 class AboutUsSettingController extends Controller
 {
     public function update(AdminAboutUsSettingUpdateRequest $request)
     {
         $aboutUsSetting = AboutUsSetting::first();
-        $storeImagePath = '';
-        $ownerImagePath = '';
+        $storeImagePath = $aboutUsSetting->store_image;
+        $ownerImagePath = $aboutUsSetting->owner_image;
         if ($request->hasFile('store_image')) {
-            if ($aboutUsSetting->store_image !== '/images/default/about-us/store_image.jpg') {
+            if ($aboutUsSetting->store_image !== '/images/default/about-us/store_image.jpg' && File::exists(public_path($aboutUsSetting->store_image)))
                 unlink(public_path($aboutUsSetting->store_image));
-            }
             $storeImage = $request->file('store_image');
-            $storeImagePath = '/images/about-us/store_image.' . $storeImage->getClientOriginalExtension();
-            $storeImage->move(public_path('images/about-us/') , 'store_image.' . $storeImage->getClientOriginalExtension());
+            $imageName = time() . '.' . $storeImage->getClientOriginalExtension();
+            $storeImagePath = '/images/about-us/' . $imageName;
+            $storeImage->move(public_path('images/about-us/'), $imageName);
         } else
             $storeImagePath = $aboutUsSetting->store_image;
         if ($request->hasFile('owner_image')) {
-            if ($aboutUsSetting->owner_image !== '/images/default/about-us/owner.jpg') {
+            if ($aboutUsSetting->owner_image !== '/images/default/about-us/owner.jpg' && File::exists(public_path($aboutUsSetting->owner_image)))
                 unlink(public_path($aboutUsSetting->owner_image));
-            }
             $ownerImage = $request->file('owner_image');
-            $ownerImagePath = '/images/about-us/owner.' . $ownerImage->getClientOriginalExtension();
-            $ownerImage->move(public_path('images/about-us/') , 'owner.' . $ownerImage->getClientOriginalExtension());
+            $imageName = time() . '.' . $ownerImage->getClientOriginalExtension();
+            $ownerImagePath = '/images/about-us/'  . $imageName;
+            $ownerImage->move(public_path('images/about-us/'),$imageName);
         } else
             $ownerImagePath = $aboutUsSetting->owner_image;
         $aboutUsSetting->update([
